refactor: add tr_torrent_metainfo::rename_path()
